requirements.txt
click==7.1.2
dominate==2.6.0
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dotenv==0.15.0
numpy==1.19.4
openpyxl==3.0.5
pandas==1.2.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
visitor==0.1.3
Werkzeug==1.0.1
xlrd==2.0.1
XlsxWriter==1.3.7
SQLAlchemy==1.3.22